agregando cambios de libro

+++ DB.sql
@@ -52,7 +52,7 @@ create table Libro(
     ID int primary key auto_increment not null,
     Titulo varchar(50) not null,
     IDAutor int not null,
-    Precio numeric not null,
+    Precio float not null,
     FOREIGN KEY (IDAutor) REFERENCES Autor(ID) ON DELETE CASCADE
 );
